refactor: Route 기능 분리
