+++ src/main/java/application/Main.java
@@ -1,5 +1,10 @@
 package application;
 
+import application.routes.Route;
+import spark.ModelAndView;
+import spark.template.handlebars.HandlebarsTemplateEngine;
+
+import java.util.HashMap;
 import java.util.Map;
 
 import static spark.Spark.*;
@@ -11,11 +16,6 @@ public class Main {
 //        gameController.run();
 
         port(8080);
-        get("/hello-world", (req, res) -> {
-            return "Hello, world!";
-        });
-
+        Route.run();
     }
-
-
 }
